+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VciClientIntegratedDPoPTest.kt
@@ -155,7 +155,6 @@ val OpenId4VciClientIntegratedDPoPTest by matrixSuite {
                         val authn = request.headers[HttpHeaders.Authorization].shouldNotBeNull()
                         credentialIssuer.credential(
                             authorizationHeader = authn,
-                            acceptHeader = request.headers[HttpHeaders.Accept].shouldNotBeNull(),
                             params = WalletService.CredentialRequest.parse(requestBody).getOrThrow(),
                             credentialDataProvider = TestUtils.credentialDataProviderFun(
                                 scheme,