+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VciClientWithEncryptionTest.kt
@@ -169,7 +169,6 @@ val OpenId4VciClientWithEncryptionTest by matrixSuite {
                     val authn = request.headers[HttpHeaders.Authorization].shouldNotBeNull()
                     credentialIssuer.credential(
                         authorizationHeader = authn,
-                        acceptHeader = request.headers[HttpHeaders.Accept].shouldNotBeNull(),
                         params = WalletService.CredentialRequest.parse(requestBody).getOrThrow(),
                         credentialDataProvider = credentialDataProviderFun(
                             scheme = scheme,