+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oauth2/AuthorizationService.kt
@@ -80,7 +80,6 @@ interface AuthorizationService {
      */
     suspend fun userInfo(
         authorizationHeader: String,
-        acceptHeader: String,
         httpRequest: RequestInfo? = null,
     ): KmmResult<JsonObject>