+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidvci/CredentialIssuer.kt
@@ -33,8 +33,6 @@ import at.asitplus.wallet.lib.oauth2.RequestInfo
 import at.asitplus.wallet.lib.oauth2.ValidatedAccessToken
 import at.asitplus.wallet.lib.oidvci.OAuth2Exception.*
 import io.github.aakira.napier.Napier
-import io.ktor.http.ContentType
-import io.ktor.http.HttpHeaders
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.decodeFromJsonElement
 import kotlin.jvm.JvmOverloads
@@ -201,12 +199,9 @@ class CredentialIssuer @JvmOverloads constructor(
         params: WalletService.CredentialRequest,
         credentialDataProvider: CredentialDataProviderFun,
         request: RequestInfo? = null,
-        acceptHeader: String = request?.headers?.get(HttpHeaders.Accept)
-            ?: ContentType.Application.Json.toString(),
     ): KmmResult<CredentialResponse> = catching {
         credentialInternal(
             authorizationHeader = authorizationHeader,
-            acceptHeader = acceptHeader,
             request = params.decryptIfNeeded(),
             credentialDataProvider = credentialDataProvider,
             requestInfo = request,
@@ -221,7 +216,6 @@ class CredentialIssuer @JvmOverloads constructor(
 
     private suspend fun credentialInternal(
         authorizationHeader: String,
-        acceptHeader: String,
         request: CredentialRequestParameters,
         credentialDataProvider: CredentialDataProviderFun,
         requestInfo: RequestInfo? = null,
@@ -232,12 +226,11 @@ class CredentialIssuer @JvmOverloads constructor(
         encryptionService.validateRequestEncryption(request, hasBeenEncrypted)
         val validated = authorizationService.validateAccessToken(
             authorizationHeader,
-            acceptHeader,
             requestInfo,
         ).getOrThrow()
         request.validateAgainstToken(validated)
         val userInfo = validated.userInfoExtended
-            ?: loadUserInfo(authorizationHeader, acceptHeader, requestInfo)
+            ?: loadUserInfo(authorizationHeader, requestInfo)
         val (scheme, representation) = request.extractCredentialRepresentation()
         val responseParameters = proofValidator.validateProofExtractSubjectPublicKeys(request).map { subjectPublicKey ->
             issuer.issueCredential(
@@ -261,11 +254,9 @@ class CredentialIssuer @JvmOverloads constructor(
 
     private suspend fun loadUserInfo(
         authorizationHeader: String,
-        acceptHeader: String,
         request: RequestInfo?,
     ): OidcUserInfoExtended = authorizationService.getUserInfo(
         authorizationHeader = authorizationHeader,
-        acceptHeader = acceptHeader,
         httpRequest = request
     ).getOrThrow().let {
         OidcUserInfoExtended.fromJsonObject(it).getOrThrow()