+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/OidvciCodeFlowTest.kt
@@ -176,7 +176,6 @@ val OidvciCodeFlowTest by matrixSuite {
 
             val credential = it.issuer.credential(
                 authorizationHeader = token.toHttpHeaderValue(),
-                acceptHeader = ContentType.Application.Json.toString(),
                 params = it.client.createCredential(
                     tokenResponse = token,
                     metadata = it.issuer.metadata,
@@ -215,7 +214,6 @@ val OidvciCodeFlowTest by matrixSuite {
             requestOptions.forEach { requestOption ->
                 it.issuer.credential(
                     authorizationHeader = token.toHttpHeaderValue(),
-                    acceptHeader = ContentType.Application.Json.toString(),
                     params = it.client.createCredential(
                         tokenResponse = token,
                         metadata = it.issuer.metadata,
@@ -291,7 +289,6 @@ val OidvciCodeFlowTest by matrixSuite {
 
             val credentials: Collection<CredentialResponseSingleCredential> = it.issuer.credential(
                 authorizationHeader = token.toHttpHeaderValue(),
-                acceptHeader = ContentType.Application.Json.toString(),
                 params = WalletService.CredentialRequest.Plain(credentialRequest),
                 credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
             ).getOrThrow()
@@ -338,7 +335,6 @@ val OidvciCodeFlowTest by matrixSuite {
 
             val credential = it.issuer.credential(
                 authorizationHeader = token.toHttpHeaderValue(),
-                acceptHeader = ContentType.Application.Json.toString(),
                 params = it.client.createCredential(
                     tokenResponse = token,
                     metadata = it.issuer.metadata,
@@ -364,7 +360,6 @@ val OidvciCodeFlowTest by matrixSuite {
 
             val credential = it.issuer.credential(
                 authorizationHeader = token.toHttpHeaderValue(),
-                acceptHeader = ContentType.Application.Json.toString(),
                 params = it.client.createCredential(
                     tokenResponse = token,
                     metadata = it.issuer.metadata,
@@ -425,7 +420,6 @@ val OidvciCodeFlowTest by matrixSuite {
 
             val credential = it.issuer.credential(
                 authorizationHeader = token.toHttpHeaderValue(),
-                acceptHeader = ContentType.Application.Json.toString(),
                 params = it.client.createCredential(
                     tokenResponse = token,
                     metadata = it.issuer.metadata,
@@ -455,7 +449,6 @@ val OidvciCodeFlowTest by matrixSuite {
 
             val credential = it.issuer.credential(
                 authorizationHeader = token.toHttpHeaderValue(),
-                acceptHeader = ContentType.Application.Json.toString(),
                 params = it.client.createCredential(
                     tokenResponse = token,
                     metadata = it.issuer.metadata,
@@ -598,7 +591,6 @@ val OidvciCodeFlowTest by matrixSuite {
             shouldThrow<OAuth2Exception> {
                 it.issuer.credential(
                     authorizationHeader = token.toHttpHeaderValue(),
-                    acceptHeader = ContentType.Application.Json.toString(),
                     params = it.client.createCredential(
                         tokenResponse = token,
                         metadata = it.issuer.metadata,
@@ -625,7 +617,6 @@ val OidvciCodeFlowTest by matrixSuite {
                 val token = it.getToken(authorizationDetails)
                 it.issuer.credential(
                     authorizationHeader = token.toHttpHeaderValue(),
-                    acceptHeader = ContentType.Application.Json.toString(),
                     params = it.client.createCredential(
                         tokenResponse = token,
                         metadata = it.issuer.metadata,
@@ -648,7 +639,6 @@ val OidvciCodeFlowTest by matrixSuite {
 
             val credential = it.issuer.credential(
                 authorizationHeader = token.toHttpHeaderValue(),
-                acceptHeader = ContentType.Application.Json.toString(),
                 params = it.client.createCredential(
                     tokenResponse = token,
                     metadata = it.issuer.metadata,